# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sistema de Gestion de Minimercado",
    "description": "Backend académico para ventas, inventario, usuarios, proveedores, órdenes de compra, facturación y reportes de un minimercado.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/categories": {
      "get": {
        "tags": [
          "Categorías"
        ],
        "summary": "List Categories",
        "operationId": "list_categories_api_v1_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryRead"
                  },
                  "type": "array",
                  "title": "Response List Categories Api V1 Categories Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Categorías"
        ],
        "summary": "Create Category",
        "operationId": "create_category_api_v1_categories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/categories/{category_id}": {
      "patch": {
        "tags": [
          "Categorías"
        ],
        "summary": "Update Category",
        "operationId": "update_category_api_v1_categories__category_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Categorías"
        ],
        "summary": "Delete Category",
        "operationId": "delete_category_api_v1_categories__category_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products": {
      "get": {
        "tags": [
          "Productos e inventario"
        ],
        "summary": "List Products",
        "operationId": "list_products_api_v1_products_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductRead"
                  },
                  "type": "array",
                  "title": "Response List Products Api V1 Products Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Productos e inventario"
        ],
        "summary": "Create Product",
        "operationId": "create_product_api_v1_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/products/low-stock": {
      "get": {
        "tags": [
          "Productos e inventario"
        ],
        "summary": "Low Stock",
        "operationId": "low_stock_api_v1_products_low_stock_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductRead"
                  },
                  "type": "array",
                  "title": "Response Low Stock Api V1 Products Low Stock Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/products/{product_id}": {
      "get": {
        "tags": [
          "Productos e inventario"
        ],
        "summary": "Get Product",
        "operationId": "get_product_api_v1_products__product_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Productos e inventario"
        ],
        "summary": "Update Product",
        "operationId": "update_product_api_v1_products__product_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/purchase-orders": {
      "get": {
        "tags": [
          "Órdenes de compra"
        ],
        "summary": "List Orders",
        "operationId": "list_orders_api_v1_purchase_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PurchaseOrderRead"
                  },
                  "type": "array",
                  "title": "Response List Orders Api V1 Purchase Orders Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Órdenes de compra"
        ],
        "summary": "Create Order",
        "operationId": "create_order_api_v1_purchase_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseOrderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/purchase-orders/{order_id}": {
      "get": {
        "tags": [
          "Órdenes de compra"
        ],
        "summary": "Get Order",
        "operationId": "get_order_api_v1_purchase_orders__order_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/purchase-orders/{order_id}/receive": {
      "post": {
        "tags": [
          "Órdenes de compra"
        ],
        "summary": "Receive Order",
        "operationId": "receive_order_api_v1_purchase_orders__order_id__receive_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/sales-summary": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Sales Summary",
        "operationId": "sales_summary_api_v1_reports_sales_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/reports/inventory-movements": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Inventory Movements",
        "operationId": "inventory_movements_api_v1_reports_inventory_movements_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InventoryMovementRead"
                  },
                  "type": "array",
                  "title": "Response Inventory Movements Api V1 Reports Inventory Movements Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "CategoryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CategoryCreate"
      },
      "CategoryRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description"
        ],
        "title": "CategoryRead"
      },
      "CategoryUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "CategoryUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InventoryMovementRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "movement_type": {
            "type": "string",
            "title": "Movement Type"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "previous_stock": {
            "type": "integer",
            "title": "Previous Stock"
          },
          "new_stock": {
            "type": "integer",
            "title": "New Stock"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "movement_type",
          "quantity",
          "date",
          "reference",
          "previous_stock",
          "new_stock",
          "product_id",
          "user_id"
        ],
        "title": "InventoryMovementRead"
      },
...
      "ProductCreate": {
        "properties": {
          "barcode": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Barcode"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "purchase_price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Purchase Price"
          },
          "sale_price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Sale Price"
          },
          "current_stock": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Current Stock",
            "default": 0
          },
          "minimum_stock": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Minimum Stock",
            "default": 0
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          }
        },
        "type": "object",
        "required": [
          "barcode",
          "name",
          "purchase_price",
          "sale_price",
          "category_id"
        ],
        "title": "ProductCreate"
      },
      "ProductRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "barcode": {
            "type": "string",
            "title": "Barcode"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "purchase_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Purchase Price"
          },
          "sale_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Sale Price"
          },
          "current_stock": {
            "type": "integer",
            "title": "Current Stock"
          },
          "minimum_stock": {
            "type": "integer",
            "title": "Minimum Stock"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "barcode",
          "name",
          "description",
          "purchase_price",
          "sale_price",
          "current_stock",
          "minimum_stock",
          "active",
          "category_id"
        ],
        "title": "ProductRead"
      },
      "ProductUpdate": {
        "properties": {
          "barcode": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Barcode"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "purchase_price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Price"
          },
          "sale_price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sale Price"
          },
          "minimum_stock": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Stock"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          }
        },
        "type": "object",
        "title": "ProductUpdate"
      },
      "PurchaseOrderCreate": {
        "properties": {
          "supplier_id": {
            "type": "integer",
            "title": "Supplier Id"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/PurchaseOrderLineCreate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "supplier_id",
          "details"
        ],
        "title": "PurchaseOrderCreate"
      },
...
      "PurchaseOrderRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "delivery_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Date"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total"
          },
          "supplier_id": {
            "type": "integer",
            "title": "Supplier Id"
          },
          "administrator_id": {
            "type": "integer",
            "title": "Administrator Id"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/PurchaseOrderDetailRead"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "delivery_date",
          "state",
          "total",
          "supplier_id",
          "administrator_id",
          "details"
        ],
        "title": "PurchaseOrderRead"
      },
...
      "SalesSummary": {
        "properties": {
          "sales_count": {
            "type": "integer",
            "title": "Sales Count"
          },
          "gross_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Gross Total"
          },
          "tax_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Tax Total"
          },
          "net_subtotal": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Net Subtotal"
          }
        },
        "type": "object",
        "required": [
          "sales_count",
          "gross_total",
          "tax_total",
          "net_subtotal"
        ],
        "title": "SalesSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}